api/requirements.txt
gunicorn==20.1.0
redis==4.1.4
werkzeug==2.0.3
alembic==1.6.5
psycopg2-binary==2.9.1
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
fastapi==0.88.0
python-multipart==0.0.5
python-jose==3.3.0
passlib==1.7.4
uvicorn==0.20.0
SQLAlchemy==1.3.24
pydantic==1.10.2
email-validator==1.3.0
pytest==7.2.0
pytest-asyncio==0.20.3
pytest-mock==3.10.0
python-dotenv==0.21.0 
httpx==0.23.1
httptools==0.5.0
ujson==5.6.0
orjson==3.8.3
jinja2==3.1.2
itsdangerous==2.1.2
python-dateutil==2.8.2
black==22.12.0
numpy==1.24.3
pandas==2.0.3
scikit-learn==1.3.0
geopy==2.3.0
pytz==2022.7 